remotive_car/models/ihu/python/pyproject.toml
[project]
name = "remotivelabs-topology-remotive-car-ihu-behavioral-model-example"
description = "RemotiveTopology Remotive Car IHU Behavioral Model example"
version = "1.0.0"

requires-python = ">=3.10,<4.0"

dependencies = [
  "structlog==25.4.0",
  "remotivelabs-broker~=0.9.1",
  "remotivelabs-topology~=0.9.4",
]

[dependency-groups]
dev = ["poethepoet>=0.18.0", "ruff>=0.11.10", "mypy>=1.14.1"]

[tool.uv.build-backend]
module-root = ""
module-name = ["ihu"]

[tool.poe.tasks]
format = [{ cmd = "ruff format ." }, { cmd = "ruff check --fix ." }]
ruff = [{ cmd = "ruff check ." }, { cmd = "ruff format --check --diff ." }]
mypy = [{ cmd = "mypy ." }]
lint = ["ruff", "mypy"]
test = { cmd = "echo no test in project" }

[tool.ruff]
extend = "../../../../ruff.toml"
exclude = ["ihu/android/libs/vhal"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_calls = false
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
hide_error_codes = false
exclude = "^ihu/android/libs/vhal(/.*)?$"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
